Modify sidebar proj del func styling.
On delBtn select, proj name with strikethrough appears in place of proj
name. Underneath is del confirm message with confirm/cancel buttons.
This is done by adding special case to delete lines in swapOutLine. The
normal function is ran with the del message as the "first elm". Then the
stylized project name is prepended on top of that (all inside the
cancel/confirm wrapper that "replaces" the sidebar project line).

Sidebar proj edit/del btns still need logic.