Add logic to dueDate ConfirmEditBtn
Edit lines stacking up horizontally bug follow up:
The "reset edit line" functionality is in the cancel btn on the edit
line. The difficulty in resetting the line when the user clicks
elsewhere is that you need a temp eventListner that removes itself after
firing.

The issue arises when the user clicks confirm with an input error;
there's not a good way for the temp eventListner to know that the
confirmBtn click was no good and it removes the EL anyway.

There are also issues with the temp EL not allowing outside clicks to
propagate. EX: if edit line it shown, users needs two clicks to hide
todo card.