Fix bug: edit line reset on bad confirm click
This fixes the bug where the edit-line would not be reset on a click off
of the confirm/cancel-line after the user clicked the confirm-button but
the confirm-logic would not fire do to an input error (document event
listener (EL) remover could only remove the reset-line-listener when
the confirmBtn was clicked in general; we needed to tie the EL remover
to confirm logic firing).

Solution was to move the confirmCancelLine EL logic to EL module and
and add a "document EL needs removal tracker" with it (global var in EL
mod). Then whereever a confirm button's logic was fired, that variable
was updated to true. And then the document EL remover coudl work as it
should.

Not sure if this is SOLID, but I feel being SOLID is what got me here.
In this isssue, JS could only remove an eventListener in the same scope
where it was created. But by coding everything to have a "single
responsibilty", the EL removal logic depended on something happending
outside that scope: checkmate.

The only solution was to have a variable both scopes could access.