Bug fix: after editBtn, could not click unhide btn
What was occuring was there was an additional event.target.click() in
the resetlistener. This was originally to solve the issue of a
new-edit-line not popping up on first click if user currently had an
edit-line already open. However, if the user tried to unhide a card
while already having an edit-line open, the unhide-btn would not trigger
on first click. The was because the extra click would click the
unhide-btn twice (thus unhiding and then rehiding the card).

The original issue was due to this: If the user clicked on an edit-btn
with a edit-line aleady open, the resetlistener is applied to that new
edit-line. Then, the old cancelClick in the first resetlistener (outside
of new editLine) would trigger and click the old cancelClick (like it
should). But then, the new resetlistener would register the old
cancelClick as an outside event and then click the new cancelBtn (as it
should). Thus making it seem like the new edit-line was not triggering
(it was, but was then immediately cancelling).

Solution was the wrap the resetlistener logic in a if/then that checks
if the new outside event was inside another cancel-btn. If it was, do
nothing (the resetlistener logic fires on the actual user event click.
Now, the internal cancelClick is being bypassed for the new
resetlistener).

With the original issue solved, the secondary "card won't unhide issue"
solves itself (no more extra event.target.click()).