Fix bug: hideDetailsBtn no work after editLines
This fix changes the mechanics behind switching from the
abridged/detailed views in the todo cards. Now, the display is changed
to none/intial for the specific elms in js, not css. This is because js
manually sets displays when the editLines are used, which overwrites
the abridged class display setting in css.

Also, wrapped all todoCard-data-lines in container (per card) to avoid
the bottom-buttons jumping up next to the priority-edit-line (edit-line
is inline).